backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
google-generativeai==0.3.2
python-dotenv==1.0.1
pydantic==2.6.4
starlette==0.36.3
python-jose==3.3.0
passlib==1.7.4
CORSHeader==0.1.0